+++ package.json
@@ -3083,6 +3083,52 @@
 		],
 		"languageModelTools": [
 			{
+				"name": "github-pull-request_fetch",
+				"tags": [
+					"github",
+					"issues"
+				],
+				"toolReferenceName": "fetch",
+				"displayName": "Get GitHub Issue/PR",
+				"modelDescription": "Get a GitHub issue/PR, which includes the title, body, and comments.",
+				"icon": "$(info)",
+				"canBeReferencedInPrompt": true,
+				"parametersSchema": {
+					"type": "object",
+					"properties": {
+						"repo": {
+							"type": "object",
+							"description": "The repository to get the issue from.",
+							"properties": {
+								"owner": {
+									"type": "string",
+									"description": "The owner of the repository to get the issue from."
+								},
+								"name": {
+									"type": "string",
+									"description": "The name of the repository to get the issue from."
+								}
+							},
+							"required": [
+								"owner",
+								"name"
+							]
+						},
+						"issueNumber": {
+							"type": "number",
+							"description": "The number of the issue to get."
+						}
+					},
+					"required": [
+						"issueNumber"
+					]
+				},
+				"supportedContentTypes": [
+					"text/plain"
+				],
+				"when": "config.githubPullRequests.experimental.chat"
+			},
+			{
 				"name": "github-pull-request_summarize",
 				"tags": [
 					"github",
